moved helpers.js content to this file and  added comments
